moved PendingFuture to core.lib
