+++ HABApp/core/const/__init__.py
@@ -2,6 +2,3 @@ from . import json
 from . import topics
 from .const import MISSING
 from .loop import loop
-
-# utilities last!
-from . import utilities
