--- HABApp/core/const/utilities/__init__.py
@@ -1,1 +1,0 @@
-from .pending_future import PendingFuture
