+++ HABApp/core/lib/__init__.py
@@ -1,1 +1,2 @@
 from .funcs import list_files
+from .pending_future import PendingFuture
