+++ HABApp/openhab/items/__init__.py
@@ -6,7 +6,7 @@ from .switch_item import SwitchItem
 from .color_item import ColorItem
 from .number_item import NumberItem
 from .datetime_item import DatetimeItem
-from .group_item import GroupItem
 from .string_item import StringItem, LocationItem, PlayerItem
-from .thing_item import Thing
 from .image_item import ImageItem
+from .group_item import GroupItem
+from .thing_item import Thing
